+++ packages/cupertino_controllers/example/ios/Runner/GeneratedPluginRegistrant.m
@@ -3,12 +3,10 @@
 //
 
 #import "GeneratedPluginRegistrant.h"
-#import <cupertino_controllers/CupertinoControllersPlugin.h>
 
 @implementation GeneratedPluginRegistrant
 
 + (void)registerWithRegistry:(NSObject<FlutterPluginRegistry>*)registry {
-  [CupertinoControllersPlugin registerWithRegistrar:[registry registrarForPlugin:@"CupertinoControllersPlugin"]];
 }
 
 @end