+++ packages/flutter_midi/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0 - 05.04.2019
+
+* Support for Desktop (Path Only)
+
+## 0.2.0 - 04.06.2019
+
+* Updating Example to be desktop aware
+
 ## 0.1.1
 
 * Added Ability to Change .sf2 File Being Played