+++ packages/persist_theme/CHANGELOG.md
@@ -1,3 +1,28 @@
+## 1.0.3 - 05.09.2019
+
+* Support for `get darkTheme` for System wide dark mode
+* Updated Example
+
+## 1.0.2 - 05.04.2019
+
+* Support for Desktop
+
+## 1.0.1 - 04.07.2019
+
+* Updated Example
+
+## 1.0.0 - 04.07.2019
+
+* Breaking Changes: removed theme class in favor of `localstorage` plugin
+* Removed `path_provider` plugin
+* Updated some theme logic
+* Adding `key` if you want to save multiple themes
+* Changing `readFromDisk` to `init`
+* Added `reset` option
+
+## 0.1.0 - 04.06.2019
+
+* Updating Example to be Desktop Aware
 ## 0.0.3
 
 * Adding Support for ThemeData overrides